refactor(package): move methods clone, with_features, without_features from Package to PackageSpecification
